+++ src/components/App/App.css
@@ -54,6 +54,6 @@
   flex-wrap: wrap;
   justify-content: space-evenly;
   align-items: space-evenly;
-  margin: 0 80px;
+  margin: 0 300px;
   padding: 0 50px;
 }
