+++ src/components/App/App.js
@@ -1,7 +1,22 @@
-import logo from './logo.svg';
 import '../App/App.css';
 import BusinessList from '../BusinessList/BusinessList';
 import SearchBar from '../SearchBar/SearchBar';
+
+import img from '../Business/Breakfast-Burritos-1700x1272.jpg';
+
+const fakeBusiness = {
+    imageSrc: img,
+    name: 'Bewitching Breakfast Burritos',
+    address: '1034 W Lime Blvd',
+    city: 'San Diego',
+    state: 'CA',
+    zipCode: '92101',
+    category: 'Mexican',
+    rating: 4.6,
+    reviewCount: 88
+};
+
+const businessArray = [fakeBusiness, fakeBusiness, fakeBusiness, fakeBusiness, fakeBusiness, fakeBusiness];
 
 function App() {
   return (
@@ -9,7 +24,7 @@ function App() {
       <h1 className="header">Ravenous</h1>
       <SearchBar />
       <div className="businessList">
-        <BusinessList />
+        <BusinessList businesses={businessArray}/>
       </div>
     </div>
   );
